odin_Worker/requirements.txt
meshtastic==2.3.13
pypubsub==4.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1